+++ vfat_test/string_helpers.h
@@ -1,0 +1,13 @@
+#ifndef __STRING_HELPERS_H
+#define __STRING_HELPERS_H
+
+#ifdef __BUILDING_TESTS
+#include <unistd.h>
+#else
+
+#endif
+
+
+size_t find_in_string(const char *buf, char c);
+
+#endif
